Phihung

+++ view/client/cart.php
@@ -21,25 +21,33 @@
                     <td><?php echo $cart->sku ?></td>
                     <td><?php echo $cart->size ?></td>
                     <td>
-                        <form class="frm-qtt" action="index.php?act=updateCartItem" method="POST" style="display: flex; align-items: center; gap: 10px;">
-                            <input class="quantity-cartItem" type="number" required min="1" name="quantity" value="<?php echo $cart->quantity ?>">
-                            <input type="hidden" name="cartItemId" value="<?php echo $cart->id ?>"
-    >
+                        <?php
+                        // Kiểm tra tồn kho
+                        $stock_available = ($cart->stock_quantity > 0) ? $cart->stock_quantity : $cart->available_stock ;
+                        if ($stock_available > 0) { ?>
+                            <form class="frm-qtt" action="index.php?act=updateCartItem" method="POST"
+                                style="display: flex; align-items: center; gap: 10px;">
+                                <input class="quantity-cartItem" type="number" required min="1" name="quantity"
+                                    value="<?php echo $cart->quantity ?>">
+                                <input type="hidden" name="cartItemId" value="<?php echo $cart->id ?>">
                                 <button class="btn-updateCartItem" type="submit" name="submit-updateCartItemClick">
                                     <i class="fa fa-refresh" aria-hidden="true"></i>
                                 </button>
-                            <a href="index.php?act=deleteCartItem&cartItemId=<?php echo $cart->id ?>"
-                                class="btn-deleteCartItem">
-                                <i class="fa-solid fa-trash-can"></i>
-                            </a>
-                        </form>
+                                <a href="index.php?act=deleteCartItem&cartItemId=<?php echo $cart->id ?>"
+                                    class="btn-deleteCartItem">
+                                    <i class="fa-solid fa-trash-can"></i>
+                                </a>
+                            </form>
+                        <?php } else { ?>
+                            <span style="color: red; font-weight: bold;">Hết hàng</span>
+                        <?php } ?>
                     </td>
                     <td><?php echo $cart->price ?></td>
                     <td style="text-align: right; padding-right: 10px;"><?php echo $cart->total_price ?></td>
 
                 </tr>
             <?php } ?>
-            
+
             <tr style="height: 40px;">
                 <th style="text-align: right; padding-right: 10px;color: black;" colspan="6">Số lượng</th>
                 <td style="text-align: right; padding-right: 10px;"><?php echo $cartAll->total_items ?></td>
